# frontend/src/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "API Zümm",
    "description" : "SIG apicole de gestion et de suivi de ruchers — contrat REST public. Les identifiants d'API (ex. getZummHoneyActualQuantity) ne se traduisent pas.",
    "license" : {
      "name" : "Propriétaire"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/ruchers/{siteId}/fiche-inspection.pdf" : {
      "get" : {
        "tags" : [ "rucher-controller" ],
        "operationId" : "ficheInspection",
        "parameters" : [ {
          "name" : "siteId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/pdf" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
    "/api/ruchers/{siteId}/emport" : {
      "get" : {
        "tags" : [ "rucher-controller" ],
        "operationId" : "emporter",
        "parameters" : [ {
          "name" : "siteId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EmportRucher"
                }
              }
            }
          }
        }
      }
    },
    "/api/ruchers/synthese" : {
      "get" : {
        "tags" : [ "rucher-controller" ],
        "operationId" : "synthese_1",
        "parameters" : [ {
          "name" : "siteId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SyntheseRucher"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/meteo" : {
      "get" : {
        "tags" : [ "meteo-controller" ],
        "operationId" : "pourSite",
        "parameters" : [ {
          "name" : "siteId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "jours",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 7
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeteoReponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "MeteoVisite" : {
        "type" : "object",
        "properties" : {
          "temperatureCelsius" : {
            "type" : "number"
          },
          "humiditePourcent" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 0
          },
          "ventKmh" : {
            "type" : "number"
          },
          "source" : {
            "type" : "string",
            "pattern" : "open-meteo|simulation|saisie"
          }
        }
      },
      "ObservationVisite" : {
        "type" : "object",
        "properties" : {
          "couvainOeufs" : {
            "type" : "boolean"
          },
          "couvainLarves" : {
            "type" : "boolean"
          },
          "couvainOpercule" : {
            "type" : "boolean"
          },
          "motifPonte" : {
            "type" : "string",
            "pattern" : "compact|lacunaire|irregulier|absent"
          },
          "reineVue" : {
            "type" : "boolean"
          },
          "cellulesRoyales" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "cellulesRoyalesCause" : {
            "type" : "string",
            "pattern" : "essaimage|supersedure|urgence"
          },
          "cadresCouvain" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 40,
            "minimum" : 0
          },
          "cadresMiel" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 40,
            "minimum" : 0
          },
          "cadresPollen" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 40,
            "minimum" : 0
          },
          "temperament" : {
            "type" : "string",
            "pattern" : "doux|normal|agressif"
          }
        }
      },
...
      "PointReleve" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "maxLength" : 40,
            "minLength" : 0
          },
          "coche" : {
            "type" : "boolean"
          },
          "niveau" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 3,
            "minimum" : 0
          }
        },
        "required" : [ "code" ]
      },
...
      "PathologieReponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "pathologie" : {
            "type" : "string"
          },
          "gravite" : {
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          }
        }
      },
      "PhotoReponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "cible" : {
            "type" : "string",
            "enum" : [ "VISITE", "RUCHE", "SITE", "REINE", "RECOLTE", "TRAITEMENT" ]
          },
          "cibleId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "url" : {
            "type" : "string"
          },
          "legende" : {
            "type" : "string"
          },
          "creeLe" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "VisiteReponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "rucheId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "rucheModele" : {
            "type" : "string"
          },
          "agentId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "agentNom" : {
            "type" : "string"
          },
          "planningId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "dateVisite" : {
            "type" : "string",
            "format" : "date"
          },
          "heureVisite" : {
            "type" : "string",
            "format" : "time",
            "example" : "14:30:00"
          },
          "dureeMin" : {
            "type" : "integer",
            "format" : "int32"
          },
          "raison" : {
            "type" : "string",
            "enum" : [ "controle", "recolte", "traitement", "nourrissage", "division", "autre" ]
          },
          "constatations" : {
            "type" : "string"
          },
          "actionsPrevues" : {
            "type" : "string"
          },
          "actionsEffectuees" : {
            "type" : "string"
          },
          "recommandations" : {
            "type" : "string"
          },
          "effectifQualitatif" : {
            "type" : "string",
            "enum" : [ "faible", "moyen", "fort" ]
          },
          "etatSante" : {
            "type" : "string",
            "enum" : [ "bon", "moyen", "mauvais" ]
          },
          "productivite" : {
            "type" : "integer",
            "format" : "int32"
          },
          "observation" : {
            "$ref" : "#/components/schemas/ObservationVisite"
          },
          "meteo" : {
            "$ref" : "#/components/schemas/MeteoVisite"
          },
          "pathologies" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PathologieReponse"
            }
          },
          "points" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PointReleve"
            }
          },
          "photos" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PhotoReponse"
            }
          },
          "creeLe" : {
            "type" : "string",
            "format" : "date-time"
          },
          "majLe" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "TacheReponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "libelle" : {
            "type" : "string"
          },
          "rucheId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "rucheModele" : {
            "type" : "string"
          },
          "agentId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "agentNom" : {
            "type" : "string"
          },
          "echeance" : {
            "type" : "string",
            "format" : "date"
          },
          "faite" : {
            "type" : "boolean"
          },
          "priorite" : {
            "type" : "string"
          },
          "categorie" : {
            "type" : "string"
          },
          "origine" : {
            "type" : "string"
          },
          "regleCode" : {
            "type" : "string"
          },
          "creeLe" : {
            "type" : "string",
            "format" : "date-time"
          },
          "majLe" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "RessourceFloraleReponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ressource" : {
            "type" : "string"
          },
          "distanceM" : {
            "type" : "integer",
            "format" : "int32"
          },
          "moisDebut" : {
            "type" : "integer",
            "format" : "int32"
          },
          "moisFin" : {
            "type" : "integer",
            "format" : "int32"
          },
          "note" : {
            "type" : "string"
          }
        }
      },
      "SiteReponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "nom" : {
            "type" : "string"
          },
          "fermeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "fermeNom" : {
            "type" : "string"
          },
          "latitude" : {
            "type" : "number"
          },
          "longitude" : {
            "type" : "number"
          },
          "altitude" : {
            "type" : "number"
          },
          "dateMiseEnOeuvre" : {
            "type" : "string",
            "format" : "date"
          },
          "dateDemenagement" : {
            "type" : "string",
            "format" : "date"
          },
          "dateCloture" : {
            "type" : "string",
            "format" : "date"
          },
          "adresseRue" : {
            "type" : "string"
          },
          "codePostal" : {
            "type" : "string"
          },
          "ville" : {
            "type" : "string"
          },
          "pays" : {
            "type" : "string"
          },
          "typeSite" : {
            "type" : "string"
          },
          "exposition" : {
            "type" : "string"
          },
          "ressources" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RessourceFloraleReponse"
            }
          },
          "priorite" : {
            "type" : "string"
          },
          "couvertureReseau" : {
            "type" : "string"
          },
          "creeLe" : {
            "type" : "string",
            "format" : "date-time"
          },
          "majLe" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "CompartimentReponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "corps", "hausse" ]
          },
          "nbCadres" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "RucheReponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "modele" : {
            "type" : "string"
          },
          "siteId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "siteNom" : {
            "type" : "string"
          },
          "fermeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "fermeNom" : {
            "type" : "string"
          },
          "agentResponsableId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "agentResponsableNom" : {
            "type" : "string"
          },
          "etat" : {
            "type" : "string",
            "enum" : [ "creee", "peuplee", "active", "en_division", "en_collecte", "cloturee" ]
          },
          "nbHausses" : {
            "type" : "integer",
            "format" : "int32"
          },
          "compartiments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CompartimentReponse"
            }
          },
          "typeRuche" : {
            "type" : "string"
          },
          "couleur" : {
            "type" : "string"
          },
          "origine" : {
            "type" : "string"
          },
          "causeCloture" : {
            "type" : "string"
          },
          "priorite" : {
            "type" : "string"
          },
          "creeLe" : {
            "type" : "string",
            "format" : "date-time"
          },
          "majLe" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "TraitementReponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "rucheId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "rucheModele" : {
            "type" : "string"
          },
          "agentId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "agentNom" : {
            "type" : "string"
          },
          "visiteId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "produit" : {
            "type" : "string"
          },
          "substanceActive" : {
            "type" : "string"
          },
          "cible" : {
            "type" : "string"
          },
          "dose" : {
            "type" : "number"
          },
          "doseUnite" : {
            "type" : "string"
          },
          "dateDebut" : {
            "type" : "string",
            "format" : "date"
          },
          "dateFin" : {
            "type" : "string",
            "format" : "date"
          },
          "delaiCarenceJours" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dateRetrait" : {
            "type" : "string",
            "format" : "date"
          },
          "sousCarence" : {
            "type" : "boolean"
          },
          "ordonnance" : {
            "type" : "string"
          },
          "ordonnanceVeterinaire" : {
            "type" : "string"
          },
          "ordonnanceDate" : {
            "type" : "string",
            "format" : "date"
          },
          "note" : {
            "type" : "string"
          },
          "creeLe" : {
            "type" : "string",
            "format" : "date-time"
          },
          "majLe" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "EmportRucher" : {
        "type" : "object",
        "properties" : {
          "preleveLe" : {
            "type" : "string",
            "format" : "date-time"
          },
          "site" : {
            "$ref" : "#/components/schemas/SiteReponse"
          },
          "ruches" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RucheReponse"
            }
          },
          "dernieresVisites" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/VisiteReponse"
            }
          },
          "tachesOuvertes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TacheReponse"
            }
          },
          "sousCarence" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TraitementReponse"
            }
          }
        }
      },
      "SyntheseRucher" : {
        "type" : "object",
        "properties" : {
          "siteId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "siteNom" : {
            "type" : "string"
          },
          "ville" : {
            "type" : "string"
          },
          "priorite" : {
            "type" : "string"
          },
          "nbRuches" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nbActives" : {
            "type" : "integer",
            "format" : "int32"
          },
          "santeMoyenne" : {
            "type" : "integer",
            "format" : "int32"
          },
          "coloniesEvaluees" : {
            "type" : "integer",
            "format" : "int32"
          },
          "risqueEssaimageMax" : {
            "type" : "integer",
            "format" : "int32"
          },
          "ruchesSousCarence" : {
            "type" : "integer",
            "format" : "int32"
          },
          "alertesOuvertes" : {
            "type" : "integer",
            "format" : "int64"
          },
          "tachesOuvertes" : {
            "type" : "integer",
            "format" : "int64"
          },
          "productionKg" : {
            "type" : "number"
          }
        }
      },
...
      "MeteoReponse" : {
        "type" : "object",
        "properties" : {
          "siteId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "latitude" : {
            "type" : "number",
            "format" : "double"
          },
          "longitude" : {
            "type" : "number",
            "format" : "double"
          },
          "temperatureCelsius" : {
            "type" : "number",
            "format" : "double"
          },
          "humiditePourcent" : {
            "type" : "integer",
            "format" : "int32"
          },
          "ventKmh" : {
            "type" : "number",
            "format" : "double"
          },
          "source" : {
            "type" : "string"
          },
          "instant" : {
            "type" : "string",
            "format" : "date-time"
          },
          "previsions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PrevisionJour"
            }
          }
        }
      },
      "PrevisionJour" : {
        "type" : "object",
        "properties" : {
          "date" : {
            "type" : "string",
            "format" : "date"
          },
          "temperatureMinCelsius" : {
            "type" : "number",
            "format" : "double"
          },
          "temperatureMaxCelsius" : {
            "type" : "number",
            "format" : "double"
          },
          "precipitationsMm" : {
            "type" : "number",
            "format" : "double"
          },
          "ventMaxKmh" : {
            "type" : "number",
            "format" : "double"
          }
        }
      },
...
    }
  }
}